+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/tests/latest/test_container_commands.py
@@ -4,7 +4,9 @@
 # --------------------------------------------------------------------------------------------
 
 import time
+import tempfile
 import unittest
+import yaml
 from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer, live_only
 
 
@@ -300,6 +302,42 @@ class AzureContainerInstanceScenarioTest(ScenarioTest):
         with self.assertRaisesRegexp(CloudError, "Azure Error: NetworkProfileNotFound"):
             self.cmd('container create -g {rg} -n {container_group_name} --image nginx '
                      '--network-profile {network_profile_name} ')
+
+    # Test export container.
+    @ResourceGroupPreparer()
+    def test_container_export(self, resource_group, resource_group_location):
+        container_group_name = self.create_random_name('clicontainer', 16)
+        image = 'nginx:latest'
+
+        _, output_file = tempfile.mkstemp()
+
+        self.kwargs.update({
+            'container_group_name': container_group_name,
+            'resource_group_location': resource_group_location,
+            'output_file': output_file,
+            'image': image,
+        })
+
+        self.cmd('container create -g {rg} -n {container_group_name} --image {image}',
+                 checks=[self.check('name', '{container_group_name}'),
+                         self.check('location', '{resource_group_location}'),
+                         self.check('provisioningState', 'Succeeded'),
+                         self.check('osType', 'Linux'),
+                         self.check('containers[0].image', '{image}'),
+                         self.exists('containers[0].resources.requests.cpu'),
+                         self.exists('containers[0].resources.requests.memoryInGb')])
+
+        self.cmd('container export -g {rg} -n {container_group_name} -f "{output_file}"')
+
+        cg_definition = None
+        with open(output_file, 'r') as f:
+            cg_definition = yaml.load(f)
+
+        self.check(cg_definition["name"], container_group_name)
+        self.check(cg_definition['properties']['containers'][0]['properties']['image'], image)
+        self.check(cg_definition['location'], resource_group_location)
+        self.check(cg_definition['properties']['containers'][0]['properties']['resources']['requests']['cpu'], 1.0)
+        self.check(cg_definition['properties']['containers'][0]['properties']['resources']['requests']['memoryInGB'], 1.5)
 
     # Test create container with azure file volume
     @ResourceGroupPreparer()